@@ -11,15 +11,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.1",
-    "@angular/common": "~13.1.1",
-    "@angular/compiler": "~13.1.1",
-    "@angular/core": "~13.1.1",
-    "@angular/forms": "~13.1.1",
-    "@angular/localize": "~13.1.1",
-    "@angular/platform-browser": "~13.1.1",
-    "@angular/platform-browser-dynamic": "~13.1.1",
-    "@angular/router": "~13.1.1",
+    "@angular/animations": "~14.2.12",
+    "@angular/common": "~14.2.12",
+    "@angular/compiler": "~14.2.12",
+    "@angular/core": "~14.2.12",
+    "@angular/forms": "~14.2.12",
+    "@angular/localize": "~14.2.12",
+    "@angular/platform-browser": "~14.2.12",
+    "@angular/platform-browser-dynamic": "~14.2.12",
+    "@angular/router": "~14.2.12",
     "@ng-bootstrap/ng-bootstrap": "^11.0.1",
     "@ngrx/effects": "^13.0.2",
     "@ngrx/router-store": "^13.0.2",
@@ -34,9 +34,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.2",
+    "@angular-devkit/build-angular": "~14.2.10",
     "@angular/cli": "^15.0.4",
-    "@angular/compiler-cli": "~13.1.1",
+    "@angular/compiler-cli": "~14.2.12",
     "@types/jasmine": "^3.10.3",
     "@types/jasminewd2": "^2.0.10",
     "@types/node": "^12.20.41",
@@ -51,6 +51,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "^6.1.3",
-    "typescript": "~4.5.4"
+    "typescript": "~4.8.4"
   }
 }
